requirements.txt
appdirs==1.4.3
astroid==1.5.3
atomicwrites==1.1.5
autopep8==1.3.2
backports.csv==1.0.5
Beaker==1.8.1
beautifulsoup4==4.6.0
bleach==2.1.1
bs4==0.0.1
certifi==2017.4.17
cffi==1.11.2
chardet==3.0.4
cli-helpers==1.0.1
click==6.7
click-default-group==1.2
click-log==0.2.1
click-threading==0.4.4
colorama==0.3.9
colour-valgrind==0.3.8
configobj==5.0.6
cupshelpers==1.0
decorator==4.1.2
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.12.2
Flask-Compress==1.4.0
Flask-Cors==3.0.3
Flask-SocketIO==2.9.2
gcalcli==3.4.0
gdbgui==0.9.1.0
gevent==1.2.2
google-api-python-client==1.6.4
greenlet==0.4.13
gufw==18.4.0
gunicorn==19.7.1
html5lib==1.0b10
httpie==0.9.9
httplib2==0.10.3
humanize==0.5.1
icalendar==4.0.1
idna==2.5
ipdb==0.10.3
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.4
isort==4.2.15
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
khal==0.9.8
lark-parser==0.5.5
lazy-object-proxy==1.3.1
louis==3.4.0
lxml==4.0.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
meson==0.45.0
mistune==0.8.1
Mopidy==2.1.0
msgpack==0.5.6
nbconvert==5.3.1
nbformat==4.4.0
neovim==0.2.4
notebook==5.2.1
npyscreen==4.10.5
oauth2client==4.1.2
olefile==0.44
packaging==16.8
pacman-mirrors==4.8.3
pandocfilters==1.4.2
parso==0.1.0
pbr==3.1.1
pexpect==4.2.1
pgcli==1.8.2
pgspecial==1.9.0
pickleshare==0.7.4
Pillow==4.2.0
prompt-toolkit==1.0.15
psutil==5.4.0
psycopg2==2.7.4
ptyprocess==0.5.2
pyalsaaudio==0.8.4
pyasn1==0.3.6
pyasn1-modules==0.1.4
pybase16-builder==0.1.4
pycairo==1.16.2
pycodestyle==2.3.1
pycparser==2.18
pycups==1.9.73
pycurl==7.43.0.1
pygdbmi==0.7.4.5
pygit2==0.26.3
Pygments==2.2.0
pygobject==3.26.1
Pykka==1.2.1
pylint==1.7.2
pyparsing==2.2.0
pypugjs==4.2.2
pystache==0.5.4
pyte==0.7.0
python-dateutil==2.6.1
python-engineio==2.0.1
python-gflags==3.1.1
python-socketio==1.8.3
pytz==2017.2
pywal==1.3.3
pyxdg==0.26
PyYAML==3.12
pyzmq==16.0.3
qrcode==6.0
qtconsole==4.3.1
ranger-fm==1.9.1
readchar==0.7
regex==2017.9.23
requests==2.17.3
requests-toolbelt==0.8.0
rsa==3.4.2
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.10.0
speedtest-cli==2.0.0
sqlparse==0.2.4
stevedore==1.27.1
syslogng==1.0
tabulate==0.8.2
team==1.0
terminado==0.6
terminaltables==3.1.0
testpath==0.3.1
thefuck==3.24
tornado==4.5.2
traitlets==4.3.2
tzlocal==1.5.1
udiskie==1.7.2
uritemplate==3.0.0
urllib3==1.21.1
urwid==2.0.1
vdirsyncer==0.16.4
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.7
wrapt==1.10.11
yapf==0.19.0
